Add extensive CUDA testing for more machines
The current CUDA-related tests in the GPGPU test plan are using a snap package that contains the CUDA runtimes. Thus it does not really test what the client will have on his machine, not does it allow to chose specific version or architecture. Right now it is only compatible with CUDA 12.8, amd64 desktops and arm server machines.
This PR adds other tests that compile what is required at runtime. They assume that both the driver and CUDA are installed. They allow to chose the sample version, as well as specific sets of features.