Use text mode to capture subprocess outputs to avoid buffering issues (bugfix) (#1878)
* Use text mode to capture subprocess outputs to avoid buffering issues

When running unit tests with the buffering option, `sys.stdout.encoding`
is set to None during test execution, which leads to problems wherever
an output is decoded using it.

Instead, this commit updates all the check_output calls to use
`universal_newlines=True`, which captures the output as text. This
avoids having to decode it later.

This fixes a problem following implementation of buffering when running
unit tests in #1863.

* Fix errors