Remove requires for dss and kubectl executables
Since we do not have setup_include at the moment, the setup jobs are run
after bootstrap --- `executable` resource would not have seen dss or
kubectl when it ran during bootstrap.

We are going to disable these requires until we get to run the setup
jobs in the setup phase.