Ensure debsums job is run before wireless tests
Because of a bug in netplan[1], the miscellanea/debsums will fail if
it's run after any of the wireless tests.

Checkbox now supports the before field for job units[2].

Since wireless/detect is a dependency for all of the wireless tests,
using it ensures the miscellanea/debsums job will be called before any
wireless tests happen.

[1]: https://bugs.launchpad.net/ubuntu/+source/netplan.io/+bug/2078759
[2]: https://github.com/canonical/checkbox/pull/1935

Fix https://warthogs.atlassian.net/browse/CHECKBOX-1916