Add suspend trigger Python script
The script mimics bin/suspend.sh, but adds a `--wait` optional argument
that can be used to wait for the device to be fully initialized.

Since it's a Python script, unit tests are added.

Also modify bin/fwts_test.py so that the main function can be called
with arguments, since this is what is done from the new
bin/suspend_trigger.py script.