Update jobs
Use the new Python script, with `--wait` option when it's the first
suspend cycle after a reboot, since this is when the device is
initializing.

Some components (WWAN modules for instance) may take up to 2 minutes to
fully initialize after a reboot, so the script waits for this.

(It also didn't make sense for a device to be suspended *riught* after
being rebooted)

Remove suspend.sh script as it's not used anymore.

Fix CHECKBOX-1918