Make networking/info_device template depend on ethernet/detect
networking/info_device__index___interface template should only be run if
ethernet/detect passed.

ethernet/detect requires the `has_ethernet_adapter` manifest, so if this
manifest entry is set to false, both tests would be skipped.

Before this patch, the following problem could happen:

- on laptop without an onboard Ethernet adapter, testing teams sometimes
plug an USB Ethernet adapter to make it easier to hook the device up in
our testing lab (to be able to remote access it, etc.)
- when doing so, Checkbox will detect this adapter as if it was an
onboard adapter
- testing teams rightfully set the `has_ethernet_adapter` manifest entry
to true, yet the networking/info_device__index___interface job is still
run for that device.

This can create unnecessary problems down the line to review test
reports in order to issue a certificate.