Add wait buffer after reboot/suspend stress tests (bugfix) (#1981)
* Add suspend trigger Python script

The script mimics bin/suspend.sh, but adds a `--wait` optional argument
that can be used to wait for the device to be fully initialized.

Since it's a Python script, unit tests are added.

Also modify bin/fwts_test.py so that the main function can be called
with arguments, since this is what is done from the new
bin/suspend_trigger.py script.

* Update jobs

Use the new Python script, with `--wait` option when it's the first
suspend cycle after a reboot, since this is when the device is
initializing.

Some components (WWAN modules for instance) may take up to 2 minutes to
fully initialize after a reboot, so the script waits for this.

(It also didn't make sense for a device to be suspended *riught* after
being rebooted)

Remove suspend.sh script as it's not used anymore.

Fix CHECKBOX-1918

* Fix unit tests for Python 3.5

* Use check_call instead of check_output

* Use script arguments instead of envvars

* Adjust unit tests accordingly

* Call Checkbox jobs with arguments

* black <3 flake8

* Fix unit tests to work with py3.5