Remove check_iwlwifi_microcode_crash from SRU test plan
As discussed with HWE/Kernel team, the check_iwlwifi_microcode_crash
checks for error messages in the kernel log, but this does not
necesarily mean there is a problem with the WiFi functionality, as the
driver can fix some of these problems on the go.

This job is still interesting to have during enablement phase.

For SRU, testing the actual functionality (via the wireless connection
jobs) is enough.

Checkbox records the journal log as part of its submission, so if
wireless functionality fails, it's always possible to review if this is
due to microcode error from the journal.

Fix RTW-543