Use glmark2 for hardware rendering tests instead of unity_support_test (BugFix) (#1677)
Use glmark2, glmark2-wayland and their es2 variants to determine whether the DUT is using hardware rendering instead of unity_support_test. It checks the output of glmark2 for a GL_RENDERER string and uses the same logic as unity_support_test to see if it's produced by a software renderer.