Don't finalize the session until resume is decided
This removes session finalization in all branches. It was added because
the start_new_sesison function starts a new session, so it fails on the
second iteration. The correct solution to this was not to spam finalize
everywhere but to split that function up into start and select_test_plan

(thanks me, I know you did this to me)