Remote assistant refactoring (new) (#2057)
* Full refactoring backported to remote_assistant.py

* Backport controller refactoring

Forgotten removed api

* Fix unit tests by renaming apis and changing whats_up response

* Refactored allowed_when and _state -> state

Decorator was moved to top level of the module, state now has a
non-private property (that logs transitions) and was changed accordingly

Minor fix: use .state and not ._state in decorator

* Renamed resume_session to prepare_resume_session

Minor: forgotten 2 state comparison to str instead of new enum

* Missing bootstrap_and_continue

* Remove outdated resume machanism

* Check that all states have a connection_strategy

* Support optional parameters of interaction

* Include test plan in session metadata

* Move functions, rename functions that continue with _continue

* Default value of resume_by_id should be empty dict

* Propagate prepare_resume_session

* Also update name in the test

* Don't finalize the session until resume is decided

This removes session finalization in all branches. It was added because
the start_new_sesison function starts a new session, so it fails on the
second iteration. The correct solution to this was not to spam finalize
everywhere but to split that function up into start and select_test_plan

(thanks me, I know you did this to me)

* Port renames and new _continue process to tests

* Also fix local changes to subcommands

* Test test plan selection function

Minor: use new continue session function

* Disable fail-fast

* Update checkbox-ng/checkbox_ng/launcher/controller.py

Co-authored-by: Fernando Bravo <39527354+fernando79513@users.noreply.github.com>

---------

Co-authored-by: Fernando Bravo <39527354+fernando79513@users.noreply.github.com>