Inherit envvars more aggressively (bugfix) (#2067)
* Add $XDG_RUNTIME_DIR for alsa tests

* Inherit envvars more aggressively

The point of inheriting them from only DISPLAY having processes was to
target user processes. This prioritizes them (as we do know the
username) but still proceeds till we run out of options. Additionally
this also calculates XDG_RUNTIME_DIR and DBUS_SESSION_BUS_ADDRESS always
if it wasn't found anywhere to (try) to make the job start anyway with a
  sane-ish value

* Python 3.5 :(

* Also test the bailout mechanism

---------

Co-authored-by: liaou3 <vincent.liao@canonical.com>