Fix ALSA-related tests on Ubuntu Core
Following #2049, ALSA-related tests stopped working when running
Checkbox on Ubuntu Core.

While the changes in this PR made sense for the classic side of the
problem (point to the ALSA config files on the system rather
older/incompatible config that might be in the snap), this does not work
on Ubuntu Core where there is no ALSA configuration at all.

Testing on Ubuntu Core, it was determined that the path to
pcm/default.conf had to be added to $ALSA_CONFIG_PATH in order for
ALSA-related programs (including the alsa_test in Checkbox base
provider) to manage to output audio through the default audio card
output.

Fix RTW-552