Refactor DSS tests using siblings (New) (#2112)
* Minor reorder of jobs

Just moving the status check job definitions higher up in the file.
More structuring coming.

* Use template to refactor existing notebook tests

* Make notebook remove job depend on what creates it

We want to skip this job if the relevant create_new_notebook... job was
skipped, so we make it depend on the same thing.

* Add kustomization to enable sharing intel gpu

See
<https://documentation.ubuntu.com/data-science-stack/latest/how-to/enable-gpus/enable-intel-gpu/#enable-the-intel-gpu-plugin>

... but instead of run a `sed` command, we do it using pyyaml

* Remove double quotes from gpu plugin versions

Wow ... didn't know that checkbox will pass the quotes too!

* Apply gpu_plugin kustomization separately

Apparently, we can't trust the order in which `kubectl -k` would apply
the kustomizations, and we apparently, we need to apply the gpu_plugin
kustomization after the other two ... apparently

* Refactor kustomization and add test for patching

* Minor spelling change to summary

* Fix requiring dss instead of kubectl

This did not impact the tests because we install kubectl in the setup
jobs, but important to catch it here if it is forgotten

* Uncomment jobs in testplan

* Move constant's definition to the top

* Fix formatting?

* Switch to using explicit siblings instead of template

* Remove the before field from siblings