udevadm: ignore small and readonly removable partitions
Removable storage test fails when it tries to run on a readonly
partition (typically, a Joliet ISO9660 one), or a partition that is
writable, but too small (the test itself creates a 100 MiB file,
therefore requires a partition with at least 100 MiB on it).

This commit uses the FSTYPE and SIZE from the lsblk command to determine
whether a device fall in this category in the udevadm parser. If so, it
is ignored by the parser.

Since linux-util 2.27 (shipping with Ubuntu 16.04), lsblk has the
ability to format its output in JSON, which is easier to parse than the
default output, so this commit uses the JSON formatting.

Fix RTW-551