Ignore some partitions in udevadm parser (bugfix) (#2154)
* udevadm: ignore small and readonly removable partitions

Removable storage test fails when it tries to run on a readonly
partition (typically, a Joliet ISO9660 one), or a partition that is
writable, but too small (the test itself creates a 100 MiB file,
therefore requires a partition with at least 100 MiB on it).

This commit uses the FSTYPE and SIZE from the lsblk command to determine
whether a device fall in this category in the udevadm parser. If so, it
is ignored by the parser.

Since linux-util 2.27 (shipping with Ubuntu 16.04), lsblk has the
ability to format its output in JSON, which is easier to parse than the
default output, so this commit uses the JSON formatting.

Fix RTW-551

* Update unit tests for the udevadm parser

- lsblk test data (*.lsblk) are reformatted to JSON data (*.lsblk.json),
similar to what `lslkb --json` would output (and the additional fstype
and size fields are added with realistic values)
- the convenience function that pulls this data is updated to parse JSON
instead of raw text
- new unit tests are added to check for real use case
(HP_ELITEBOOK_835_13_INCH_G10_RECOVERY_USB_CONNECTED, which is a laptop
where a Recovery USB disk is connected – this is the origin of the
problem that is being fixed here)
- unit tests for the new lsblk-related functions are added

* Fix py3.5 test run

* fix flake8

* Dropping patch/mock from unit test