fix reading module parameters in net_driver_info
Some modules (for instance the rtl8822ce) have parameters that are not readable as text.
For instance, the rtw_country_code parameter of the rtl8822ce contains 0xff.
This prevents checkbox from giving the entire log when an issue happens because
net_driver_info will stop right after trying to read this file.