Update disk_info to use the JSON output of lsblk
This script is calling find_pkname_is_root_mountpoint, which has
recently been modified to use the JSON output of lsblk. It is therefore
modified to fit the same model.

A simple unit test is added, since there was none previously (!). It is
using contextlib.redirect_stdout() to check the printed output, as
explained in this article:

https://adamj.eu/tech/2025/08/29/python-unittest-capture-stdout-stderr/