Fix scripts relying on find_pkname_is_root_mountpoint (bugfix) (#2179)
* Update disk_info to use the JSON output of lsblk

This script is calling find_pkname_is_root_mountpoint, which has
recently been modified to use the JSON output of lsblk. It is therefore
modified to fit the same model.

A simple unit test is added, since there was none previously (!). It is
using contextlib.redirect_stdout() to check the printed output, as
explained in this article:

https://adamj.eu/tech/2025/08/29/python-unittest-capture-stdout-stderr/

* Update removable_storage_test.py to use JSON output from lsblk

* Fix flake8

* Fix importing issue when running unit tests

Due to how unit tests import packages at test time, the following error
shows up when running `./manage.py test -u` on the base provider:

ImportError: No module named 'dbus.exceptions'; 'dbus' is not a package

Since this module is not required for the tests, it is mocked to allow
the tests to run.

* Address Massi's comments

* black